pyproject.toml
[project]
name = "mbtistats-bot"
version = "0.1.0"
description = "MBTI Stats Bot Development Environment"
readme = "README.md"
requires-python = "==3.13.*"
dependencies = [
    "jinja2>=3.1.6",
    "nonebot-adapter-console>=0.9.0",
    "nonebot-adapter-onebot>=2.4.6",
    "nonebot-adapter-qq>=1.6.6",
    "nonebot-plugin-analysis-bilibili>=2.8.1",
    "nonebot-plugin-questionmark>=0.4.1",
    "nonebot2[fastapi,httpx,websockets]>=2.4.4",
    "nonebot-plugin-mbtistats",
    "openai>=1.0.0",
    "playwright>=1.56.0",
]


[tool.uv.sources]
nonebot-plugin-mbtistats = { path = "./dev-plugins/mbtistats", editable = true }


[tool.nonebot]
builtin_plugins = ["echo", "single_session"]


[tool.nonebot.adapters]
nonebot-adapter-qq = [
    { name = "QQ", module_name = "nonebot.adapters.qq" }
]
nonebot-adapter-console = [
    { name = "Console", module_name = "nonebot.adapters.console" }
]
nonebot-adapter-onebot = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
"@local" = []


[tool.nonebot.plugins]
"@local" = ["nonebot_plugin_mbtistats"]
"nonebot-plugin-analysis-bilibili" = ["nonebot_plugin_analysis_bilibili"]
"nonebot-plugin-questionmark" = ["nonebot_plugin_questionmark"]